fix: stabilize ai chat billing access state
